Cut out the function of Kernel density estimation into another script
