+++ R/RcppExports.R
@@ -1,12 +1,12 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-mode_kde <- function(points, T = 512L) {
-    .Call(`_neggfl_mode_kde`, points, T)
-}
-
 negfl_bas <- function(y, X, hparams, B) {
     .Call(`_neggfl_negfl_bas`, y, X, hparams, B)
+}
+
+mode_kde <- function(points, T = 512L) {
+    .Call(`_neggfl_mode_kde`, points, T)
 }
 
 Int_simt_log_D <- function(x, lamb, a) {
